+++ plugins/pwa/plugin-pwa/src/client/components/PwaFoundPopup.ts
@@ -1,3 +1,4 @@
+import type { Slot } from '@vuepress/helper/client'
 import { useLocale } from '@vuepress/helper/client'
 import type { PropType, SlotsType, VNode } from 'vue'
 import { Transition, defineComponent, h, onMounted, ref } from 'vue'
@@ -20,10 +21,7 @@ export const PwaFoundPopup = defineComponent({
   },
 
   slots: Object as SlotsType<{
-    default?: (props: {
-      found: boolean
-      refresh: () => void
-    }) => VNode | VNode[] | null
+    default?: Slot<{ found: boolean; refresh: () => void }>
   }>,
 
   setup(props, { slots }) {