+++ plugins/pwa/plugin-pwa/src/client/components/PwaReadyPopup.ts
@@ -1,3 +1,4 @@
+import type { Slot } from '@vuepress/helper/client'
 import { useLocale } from '@vuepress/helper/client'
 import type { PropType, SlotsType, VNode } from 'vue'
 import {
@@ -28,10 +29,7 @@ export const PwaReadyPopup = defineComponent({
   },
 
   slots: Object as SlotsType<{
-    default?: (props: {
-      isReady: boolean
-      reload: () => void
-    }) => VNode | VNode[] | null
+    default?: Slot<{ isReady: boolean; reload: () => void }>
   }>,
 
   setup(props, { slots }) {