+++ themes/theme-default/src/client/components/VPFadeSlideYTransition.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import type { VNode } from 'vue'
+import type { SingleSlot } from '@vuepress/helper/client'
 import '@vuepress/helper/transition/fade-in-down.css'
 
 defineEmits<{
@@ -7,7 +7,7 @@ defineEmits<{
   beforeLeave: []
 }>()
 defineSlots<{
-  default: () => VNode | string | null
+  default: SingleSlot
 }>()
 </script>