+++ themes/theme-default/src/client/components/global/Badge.vue
@@ -1,30 +1,27 @@
 <script setup lang="ts">
-import type { VNode } from 'vue'
+import type { Slot } from '@vuepress/helper/client'
 
-withDefaults(
-  defineProps<{
-    /**
-     * The type of the badge
-     */
-    type?: string
-    /**
-     * The text of the badge
-     */
-    text?: string
-    /**
-     * The vertical align of the badge
-     */
-    vertical?: string
-  }>(),
-  {
-    type: 'tip',
-    text: '',
-    vertical: undefined,
-  },
-)
+const {
+  type = 'tip',
+  text = '',
+  vertical = '',
+} = defineProps<{
+  /**
+   * The type of the badge
+   */
+  type?: string
+  /**
+   * The text of the badge
+   */
+  text?: string
+  /**
+   * The vertical align of the badge
+   */
+  vertical?: string
+}>()
 
 defineSlots<{
-  default?: () => VNode | VNode[] | string | null
+  default?: Slot
 }>()
 </script>
 
@@ -32,9 +29,7 @@ defineSlots<{
   <span
     class="vp-badge"
     :class="type"
-    :style="{
-      verticalAlign: vertical,
-    }"
+    :style="vertical ? { verticalAlign: vertical } : ''"
   >
     <slot>{{ text }}</slot>
   </span>